+++ hc/api/tests/test_resume.py
@@ -1,34 +1,43 @@
-from datetime import timedelta as td
-
-from django.utils.timezone import now
 from hc.api.models import Check, Flip
 from hc.test import BaseTestCase
 
 
-class PauseTestCase(BaseTestCase):
+class ResumeTestCase(BaseTestCase):
     def setUp(self):
         super().setUp()
 
-        self.check = Check.objects.create(project=self.project, status="up")
-        self.url = f"/api/v1/checks/{self.check.code}/pause"
+        self.check = Check.objects.create(project=self.project, status="paused")
+        self.url = f"/api/v1/checks/{self.check.code}/resume"
 
     def test_it_works(self):
         r = self.csrf_client.post(
             self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
+
         self.assertEqual(r.status_code, 200)
         self.assertEqual(r["Access-Control-Allow-Origin"], "*")
 
         self.check.refresh_from_db()
-        self.assertEqual(self.check.status, "paused")
+        self.assertEqual(self.check.status, "new")
+        self.assertIsNone(self.check.last_start)
+        self.assertIsNone(self.check.last_ping)
+        self.assertIsNone(self.check.alert_after)
 
-        # It should also create a Flip object, needed for accurate downtime
-        # tracking in Check.downtimes():
         flip = Flip.objects.get()
-        self.assertEqual(flip.old_status, "up")
-        self.assertEqual(flip.new_status, "paused")
+        self.assertEqual(flip.old_status, "paused")
+        self.assertEqual(flip.new_status, "new")
         # should be marked as processed from the beginning, so sendalerts ignores it
         self.assertTrue(flip.processed)
+
+    def test_it_handles_not_paused_checks(self):
+        self.check.status = "up"
+        self.check.save()
+
+        r = self.csrf_client.post(
+            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
+        )
+
+        self.assertEqual(r.status_code, 409)
 
     def test_it_accepts_api_key_in_post_body(self):
         payload = {"api_key": "X" * 32}
@@ -36,7 +45,7 @@ class PauseTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 200)
 
         self.check.refresh_from_db()
-        self.assertEqual(self.check.status, "paused")
+        self.assertEqual(self.check.status, "new")
 
     def test_it_handles_options(self):
         r = self.client.options(self.url)
@@ -49,8 +58,7 @@ class PauseTestCase(BaseTestCase):
 
     def test_it_validates_ownership(self):
         check = Check.objects.create(project=self.bobs_project, status="up")
-
-        url = f"/api/v1/checks/{check.code}/pause"
+        url = f"/api/v1/checks/{check.code}/resume"
         r = self.client.post(
             url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
@@ -58,7 +66,7 @@ class PauseTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 403)
 
     def test_it_validates_uuid(self):
-        url = "/api/v1/checks/not-uuid/pause"
+        url = "/api/v1/checks/not-uuid/resume"
         r = self.client.post(
             url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
@@ -66,40 +74,12 @@ class PauseTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 404)
 
     def test_it_handles_missing_check(self):
-        url = "/api/v1/checks/07c2f548-9850-4b27-af5d-6c9dc157ec02/pause"
+        url = "/api/v1/checks/07c2f548-9850-4b27-af5d-6c9dc157ec02/resume"
         r = self.client.post(
             url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
         )
 
         self.assertEqual(r.status_code, 404)
-
-    def test_it_clears_last_start_alert_after(self):
-        self.check.last_start = now()
-        self.check.alert_after = self.check.last_start + td(hours=1)
-        self.check.save()
-
-        r = self.client.post(
-            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
-        )
-
-        self.assertEqual(r.status_code, 200)
-        self.assertEqual(r["Access-Control-Allow-Origin"], "*")
-
-        self.check.refresh_from_db()
-        self.assertEqual(self.check.last_start, None)
-        self.assertEqual(self.check.alert_after, None)
-
-    def test_it_clears_next_nag_date(self):
-        self.profile.nag_period = td(hours=1)
-        self.profile.next_nag_date = now() + td(minutes=30)
-        self.profile.save()
-
-        self.client.post(
-            self.url, "", content_type="application/json", HTTP_X_API_KEY="X" * 32
-        )
-
-        self.profile.refresh_from_db()
-        self.assertIsNone(self.profile.next_nag_date)
 
     def test_it_rejects_non_dict_post_body(self):
         r = self.csrf_client.post(self.url, "123", content_type="application/json")